% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{OysterPop}
\alias{OysterPop}
\title{Relative abundance indecies}
\usage{
OysterPop(
  d,
  global.line = c("lm", "loess", "both"),
  river.line = c("lm", "loess", "both")
)
}
\arguments{
\item{d}{data.frame}
}
\description{
Generate abundance plots and datasets
}
